Feat/Add e2e test with Playwright (#161)
# Description
Part of #157 

Add Playwrigh and add the first few tests. The main goals are:

- [x] Add Playwright and make it work locally
- [x] Add a few tests and confirm they are working correctly (with
localhost)
- [x] Support connection to Metamask
- [x] Add a test with API mocks, specifically for subgraph
- [x] Add a test with Odoo and Wallet connection
- [x] Make the API mocks typesafe
- [x] Add documentation

⚠️ There is an issue with how the wallet is handled after being cached
for the e2e test: https://github.com/Synthetixio/synpress/issues/1103

### A few notes:

I started mocking with MSW but it was a painful experience. First,
Playwright doesn't work well with libraries that [override their service
worker
](https://playwright.dev/docs/network#missing-network-events-and-service-workers).
Then, the most promising plugin to support graphql-codegen with MSW and
Playwright does not support multiple endpoints, which is an issue for
testing the legacy flow (like the resolutions).

In the end, I'm fairly happy with the final solution. It's not a lot of
code and it doesn't use new libraries but reuses what we have without
sacrificing functionalities.

---------

Co-authored-by: Andrea Tosatto <andrea.tosatto@infinitaslearning.com>