Fix/Make it build by excluding TS errors (#166)
# Description

Closes #167

We currently let the app build with Typescript errors, and this may
cause us to deploy easy-to-catch bugs. In this PR I upgraded TS to the
latest version and added a few `// @ts-expect-error`. Additionally, I
changed the Next config to stop building on TS errors.

This PR can be merged as it is, but would be great to fix comments
before merging it.

# What to do
- [x] Be sure the app is still building on the CI
- [x] Ask Ste and Gian to check the errors and instruction on how to fix
the errors